MINITALK FINI

+++ libft/srcs/ft_atoi.c
@@ -20,7 +20,7 @@ static int	ft_isspace(char c)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int			i;
 	int			sign;
